Update slow_algorithm.py
Name : AISHWARYA B
CLG    : SRI SAIRAM INSTITUTE OF TECHNOLOGY
ID       : SIT23CS120

The code was changed to generate the entire Fibonacci sequence up to the nth term instead of calculating only a single Fibonacci number. It handles edge cases (n <= 0 or n == 1) and uses an efficient iterative approach with a list to store and compute the sequence.